Decrease default retries to fail faster
The default retries limit was 10 which is very high for a library used
for internal retires. This is basically causing friendly fire DDoS
issues an resulting in a very poor user experience as it takes ages to
fail. Reducing the default means that we should fail faster, but still
allows for a small failure tolerance.